Create SampleAdReportingDatabase.md
This is a simple representation of an AdReporting database I've built in SQLite using tables designated "accounts", "campaigns", "adtype", and "ad_performance". This is not about the actual numbers (which I made up) so much as it is about giving you an idea of my ability to sift through relevant data by building more complex SQL queries. 

I'll first create the tables and then try to pull some meaningful information out of them while leveraging a number of different SQL functions.